updated with steves comments
