+++ ion/services/mi/instrument_driver.py
@@ -15,13 +15,12 @@ __license__ = 'Apache 2.0'
 from ion.services.mi.common import BaseEnum
 from ion.services.mi.exceptions import InstrumentConnectionException 
 from ion.services.mi.common import DEFAULT_TIMEOUT
-from pyon.util.fsm import FSM
 
 class DriverChannel(BaseEnum):
     """Common channels for all sensors. Driver subclasses contain a subset."""
     INSTRUMENT = 'CHANNEL_INSTRUMENT'
+    CTD = 'CHANNEL_CTD'
     ALL = 'CHANNEL_ALL'
-
 
 class DriverCommand(BaseEnum):
     """Common driver commands
@@ -57,6 +56,7 @@ class DriverState(BaseEnum):
     AUTOSAMPLE = 'DRIVER_STATE_AUTOSAMPLE'
     TEST = 'DRIVER_STATE_TEST'
     CALIBRATE = 'DRIVER_STATE_CALIBRATE'
+    DETACHED = 'DRIVER_STATE_DETACHED'
     COMMAND = 'DRIVER_STATE_COMMAND'
 
 class DriverEvent(BaseEnum):
@@ -90,6 +90,8 @@ class DriverEvent(BaseEnum):
     RESET = 'DRIVER_EVENT_RESET'
     ENTER = 'DRIVER_EVENT_ENTER'
     EXIT = 'DRIVER_EVENT_EXIT'
+    ATTACH = 'DRIVER_EVENT_ATTACH'
+    DETACH = 'DRIVER_EVENT_DETACH'
     
 
 class DriverStatus(BaseEnum):
@@ -120,9 +122,11 @@ class ObservatoryState(BaseEnum):
     ACQUIRING = 'OBSERVATORY_STATUS_ACQUIRING'
     UNKNOWN = 'OBSERVATORY_STATUS_UNKNOWN'
 
-############
-# End states
-############
+"""
+TODO:
+Do we want timeouts in the driver interface. timeout=DEFAULT_TIMEOUT.
+How would we provide such behavior?
+"""
 
 class InstrumentDriver(object):
     """The base instrument driver class
@@ -136,40 +140,50 @@ class InstrumentDriver(object):
     """
 
     def __init__(self):
-        # Setup instance variables with instrument-specific instances.
-        # Some may be fed from the instrument protocol subclass.
+
+        # Instrument channel dict.
+        # A dictionary of channel-name keys and channel protocol object values.
+        # We need to change this to protocol or connection name, rather than channel.
+        self.channels = {}
         
-        self.instrument_connection = None
-        """An object for manipulating connect and disconnect to an instrument"""
-    
-        self.instrument_protocol = None
-        """The instrument-specific protocol object"""
-    
-        self.instrument_comms_method = None
-        """The communications method formatting object"""
-    
-        self.instrument_commands = None
-        """The instrument-specific command list"""
-    
-        self.instrument_metadata_parameters = None
-        """The instrument-specific metadata parameter list"""
-    
-        self.instrument_parameters = None
-        """The instrument-specific parameter list"""
-    
-        self.instrument_channels = None
-        """The instrument-specific channel list"""
-    
-        self.instrument_errors = None
-        """The instrument-specific error list"""
-    
-        self.instrument_capabilities = None
-        """The instrument-specific capabilities list"""
-    
-        self.instrument_status = None
-        """The instrument-specific status list"""
-    
+        # Below are old members with comments from EH.
+        #
+        # Protocol will create and own the connection it fronts.
+        #self.instrument_connection = None
+        #"""An object for manipulating connect and disconnect to an instrument"""
+        # This is the self.channels member above. Change name. A dict of protocols.
+        #self.instrument_protocol = None
+        #"""The instrument-specific protocol object"""
+        # This is supplied by the driver process that contains and creates the driver.
+        #self.instrument_comms_method = None
+        #"""The communications method formatting object"""
+        # Probably an enum class with execute commands that are possible.
+        #self.instrument_commands = None
+        #"""The instrument-specific command list"""
+        # Ditch this.
+        #self.instrument_metadata_parameters = None
+        #"""The instrument-specific metadata parameter list"""
+        # To be added. A dictionary or class that knows how to match, parse and format all of its params.
+        #self.instrument_parameters = None
+        #"""The instrument-specific parameter list"""
+        # TBD.
+        #self.instrument_channels = None
+        #"""The instrument-specific channel list"""
+        # Why bother, make errors draw from the common list.
+        #self.instrument_errors = None
+        #"""The instrument-specific error list"""
+        # This should return the instrument commands enum values
+        #self.instrument_capabilities = None
+        #"""The instrument-specific capabilities list"""
+        # TBD.
+        #self.instrument_status = None
+        #"""The instrument-specific status list"""
+        
+        # Do we need this also? Just use simple harmonizing logic
+        # at driver level if there are multiple connections. If one connection,
+        # one state machine.
         # Setup the state machine
+        """
         self.driver_fsm = FSM(DriverState.UNCONFIGURED)
         self.driver_fsm.add_transition(DriverEvent.CONFIGURE,
                                        DriverState.UNCONFIGURED,
@@ -251,145 +265,151 @@ class InstrumentDriver(object):
 
         self.driver_fsm.add_transition_catch(DriverEvent.RESET,
                                        action=self._handle_reset,
-                                       next_state=DriverState.UNCONFIGURED)
+                                       next_state=DriverState.UNCONFIGURED)    
+        """
+        
+    ########################################################################
+    # Channel connection interface.
+    ########################################################################
+    
+    def initialize(self, channels, timeout=10):
+        """
+        Return a device channel to an unconnected, unconfigured state.
+        @param channels List of channel names to initialize.
+        @param timeout Number of seconds before this operation times out
+        """
+        pass
 
-    def configure(self, params={}, timeout=DEFAULT_TIMEOUT):
-        """Configure the driver's parameters
+    def configure(self, configs, timeout=10):
+        """
+        Configure the driver for communications with an instrument channel.
+        @param config A dict containing channel name keys, with
+        dict values containing the comms configuration for the named channel.
+        @param timeout Number of seconds before this operation times out        
+        """
+        pass        
         
-        Some parameters are needed soley by the driver to interact with an
-        instrument. These parameters can be set here by the instantiating
-        class, likely based on some registry entries for the individual
-        instrument being interacted with 
-        
-        @param params A dictionary of the parameters for configuring the driver
+    def connect(self, channels, timeout=10):
+        """
+        Establish communications with a device channel.
+        @param channels List of channel names to connect.
         @param timeout Number of seconds before this operation times out
         """
-        assert(isinstance(params, dict))
-        
-    def initialize(self, timeout=DEFAULT_TIMEOUT):
+        pass
+    
+    def disconnect(self, channels, timeout=10):
         """
-        """
-        
-    def connect(self, timeout=DEFAULT_TIMEOUT):
-        """ Connect to the device
+        Disconnect communications with a device channel.
+        @param channels List of channel names to disconnect.
         @param timeout Number of seconds before this operation times out
-        @retval result Success/failure result
-        @throws InstrumentConnectionException
-        @todo determine result if already connected
+        
         """
-        # Something like self.InstrumentConnection.connect(), then set state
-    
-    def disconnect(self, timeout=DEFAULT_TIMEOUT):
+        pass
+
+    def detach(self, channels, timeout=10):
+        """
+        Disconnect communications with a device channel.
+        @param channels List of channel names to disconnect.
+        @param timeout Number of seconds before this operation times out
+        """
+        pass
+
+    ########################################################################
+    # Channel command interface.
+    ########################################################################
+
+    def get(self, params, timeout=10):
         """
         @param timeout Number of seconds before this operation times out
-        @retval result Success/failure result
-        @throws InstrumentConnectionException
-        @todo determine result if already disconnected
         """
-        # Something like self.InstrumentConnection.disconnect(), then set state
-           
-           
-    def execute(self, channels, command, timeout):
-        """
-        """
+        pass
     
-    def execute_direct(self, bytes, timeout):
+    def set(self, params, timeout=10):
         """
+        @param timeout Number of seconds before this operation times out
         """
+        pass
+
+    def execute(self, channels, command, timeout=10):
+        """
+        @param timeout Number of seconds before this operation times out
+        """
+        pass
     
-    def get(self, params, timeout):
+    def execute_direct(self, channels, bytes):
         """
         """
-        
-    def set(self, params, timeout):
-        """
-        """
-        
-    def get_metadata(self, params, timeout):
-        """
-        """
-        
-    def get_status(self, params, timeout):
-        """
-        """
-        
-    def get_capabilities(self, params, timeout):
-        """
-        """
+        pass
     
+    ########################################################################
+    # TBD.
+    ########################################################################    
+    
+    def get_status(self, params, timeout=10):
+        """
+        @param timeout Number of seconds before this operation times out
+        """
+        pass
+    
+    def get_capabilities(self, params, timeout=10):
+        """
+        @param timeout Number of seconds before this operation times out
+        """
+        pass
+
     #######################
     # State change handlers
     #######################
-    def _handle_configure(self):
-        """State change handler"""
-    
-    def _handle_initialize(self):
-        """State change handler"""
-    
-    def _handle_disconnect_failure(self):
-        """State change handler"""
-
-    def _handle_disconnect_success(self):
-        """State change handler"""
-
-    def _handle_disconnect(self):
-        """State change handler"""
-
-    def _handle_connect(self):
-        """State change handler"""
-
-    def _handle_connect_success(self):
-        """State change handler"""
-
-    def _handle_connect_failed(self):
-        """State change handler"""
-
-    def _handle_start_autosample(self):
-        """State change handler"""
-
-    def _handle_stop_autosample(self):
-        """State change handler"""
-    
-    def _handle_connection_lost(self):
-        """State change handler"""
-    
-    def _handle_reset(self):
-        """State change handler"""
-
-    def _handle_handle_data_received(self):
-        """State change handler"""
-        
-    def _handle_handle_get(self):
-        """State change handler"""
-        
-    def _handle_handle_set(self):
-        """State change handler"""
-    
-    def _handle_handle_acquire_sample(self):
-        """State change handler"""
-        
-    def _handle_handle_test(self):
-        """State change handler"""
-    
-    def _handle_handle_calibrate(self):
-        """State change handler"""  
-        
-    ##################
-    # Resource methods
-    ##################
-    def rcmd_initialize(self):
-        """
-        """
-    
-    def rcmd_configure(self, params):
-        """
-        """
-        
-    def rcmd_connect(self, timeout):
-        """
-        """
-        
-    def rcmd_disconnect(self, timeout):
-        """
-        """
-    
+    #def _handle_configure(self):
+    #    """State change handler"""
+    #
+    #def _handle_initialize(self):
+    #    """State change handler"""
+    #
+    #def _handle_disconnect_failure(self):
+    #    """State change handler"""
+    #
+    #def _handle_disconnect_success(self):
+    #    """State change handler"""
+    #
+    #def _handle_disconnect(self):
+    #    """State change handler"""
+    #
+    #def _handle_connect(self):
+    #    """State change handler"""
+    #
+    #def _handle_connect_success(self):
+    #    """State change handler"""
+    #
+    #def _handle_connect_failed(self):
+    #    """State change handler"""
+    #
+    #def _handle_start_autosample(self):
+    #    """State change handler"""
+    #
+    #def _handle_stop_autosample(self):
+    #    """State change handler"""
+    #
+    #def _handle_connection_lost(self):
+    #    """State change handler"""
+    #
+    #def _handle_reset(self):
+    #    """State change handler"""
+    #
+    #def _handle_handle_data_received(self):
+    #    """State change handler"""
+    #    
+    #def _handle_handle_get(self):
+    #    """State change handler"""
+    #    
+    #def _handle_handle_set(self):
+    #    """State change handler"""
+    #
+    #def _handle_handle_acquire_sample(self):
+    #    """State change handler"""
+    #    
+    #def _handle_handle_test(self):
+    #    """State change handler"""
+    #
+    #def _handle_handle_calibrate(self):
+    #    """State change handler"""
